rename tools directory
