+++ larrecodnn/NuGraph/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_subdirectory(NuGraphTools)
+add_subdirectory(Tools)
 
 cet_set_compiler_flags(DIAGS VIGILANT DWARF_VER 4
   WERROR
